Implement Chatroom Window
