+++ src/Interfaces/Models/Chat.d.ts
@@ -1,0 +1,7 @@
+import { ChatMember } from './ChatMember';
+
+export interface Chat {
+  id: string;
+  members?: ChatMember[];
+  eventId?: number;
+}
