+++ src/Interfaces/Models/ChatMessage.d.ts
@@ -1,0 +1,10 @@
+import { Client } from './Client';
+
+export interface ChatMessage {
+  id: string;
+  text: string;
+  chatId: string;
+  authorId: number;
+  author?: Client;
+  createdAt: string;
+}
