+++ src/Redux/Reducers/chatrooms/setActiveChatroom.ts
@@ -1,0 +1,10 @@
+import { ChatroomsState, ChatroomAction } from '@Interfaces';
+
+const setActiveChatroom = (state: ChatroomsState, action: ChatroomAction) => {
+  return {
+    ...state,
+    activeChatroom: action.chatId || '',
+  };
+};
+
+export default setActiveChatroom;
